Adds Player functiononailty
Changed the reticule that was given, shoots raycast from center of screen and only hits IF it collides with AI layer or Barrier layer.  Able to damage the barriers that are set around the scene, and break them.  After 3 seconds they will respawn and their health will continuously go up until they are full.  Adds in a bullet spark particle effect that registers when the player has hit a bot or the barrier.